feat: add details page on the right side
