+++ src/pages/_app.tsx
@@ -3,9 +3,7 @@ import type { AppProps } from 'next/app'
 import Layout from '../components/Layout'
 
 const App = ({ Component, pageProps }: AppProps) => (
-  <Layout>
-    <Component {...pageProps} />
-  </Layout>
+  <Component {...pageProps} />
 );
 
 export default App;
