+++ src/styles/CardsList.module.css
@@ -61,6 +61,11 @@
   z-index: 200;
 }
 
+.detalsPage {
+  display: flex;
+  flex-direction: row;
+}
+
 .itemDetailsBtn {
   cursor: pointer;
 }
